RavenDB-26295 - Split pull heartbeat cursor handling by direction
